requirements.txt
numpy~=1.20
environs~=9.3
discord.py~=1.6
seaborn~=0.11
aiohttp~=3.7
